feat: finalizando ideia geral da api e removendo projetos duplicados

+++ challenges/fmarca/fmarca/src/main/resources/db/migration/V2__create-table-vehicles.sql
@@ -4,7 +4,7 @@ CREATE TABLE vehicles(
   model VARCHAR(255) NOT NULL,
   color VARCHAR(255) NOT NULL,
   plate VARCHAR(255) NOT NULL,
-  occupying TINYINT(1) NOT NULL,
+  occupying INT NOT NULL,
   type VARCHAR(255) NOT NULL,
   establishment_id BIGINT NOT NULL,
   FOREIGN KEY (establishment_id) REFERENCES establishment (id)
